% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessOpenSensorWeb.R
\name{availablesensors}
\alias{availablesensors}
\title{availablesensors}
\usage{
availablesensors(url, my.device)
}
\arguments{
\item{url}{the url adress (given without "/" at the end) of a specivic device to create a GET-request for the OpensSensorweb REST API character}
}
\value{
a json file with all available sensors for the given device
}
\description{
This function lists all available sensors for a specific device on open sensor web
}
\examples{
opensensorwebr::availablesensors("https://api.sensoto.io/v1/organizations/open/networks/AMMS_WETTERDATEN", my.devices = "S034")
}
