% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessOpenSensorWeb.R
\name{etmodeldata}
\alias{etmodeldata}
\title{etmodeldata}
\usage{
etmodeldata(
  url = "https://api.opensensorweb.de/v0/networks/AMMS_WETTERDATEN",
  my.device = "S021",
  my.startdate = "2015-09-18T00:00:00Z",
  my.interval = 1000,
  ID.GlobRad = "Globalstrahlg_200cm",
  ID.AirTemp = "Lufttemp_200cm",
  ID.RH = "Luftfeuchtigkeit_200cm",
  ID.Rain = "Niederschlag",
  ID.Wind = "Windgeschw_250cm",
  file = "temp/Wetter_Coswig_",
  write.RData = FALSE,
  write.csv = FALSE
)
}
\arguments{
\item{url}{the url of the device in form of https://api.opensensorweb.de/v0/networks/AMMS_WETTERDATEN/devices/S034}
}
\value{
a json file with all available sensors
}
\description{
Provide all a dataset usefull for evapotranspiration modelling if all data are available at one station. Data required at the moment is: GlobalRadiation, Air Temperature, Relative Humidity, Windspeed, Rain
Todo: make Rain optional!
}
\examples{
availablesensors(https://api.opensensorweb.de/v0/networks/AMMS_WETTERDATEN/devices/S034)
}
