% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessOpenSensorWeb.R
\name{hourly}
\alias{hourly}
\title{hourly}
\usage{
hourly(
  url = "https://api.sensoto.io/v1/organizations/open/networks/AMMS_WETTERDATEN/devices",
  my.device,
  my.sensor,
  my.interval = 1000,
  my.startdate = "2015-09-18T00:00:00Z",
  aggregation = "MEAN"
)
}
\arguments{
\item{url}{the url adress (given without "/" at the end) of a specivic device to create a GET-request for the OpensSensorweb REST API character ()}

\item{my.device}{the device ID in the form "S021", can be defined by searches on https://www.opensensorweb.de}

\item{my.sensor}{the sensor ID in the form "Niederschlag", can be chosen by searches on https://www.opensensorweb.de or an request using availablesensors()}

\item{my.interval}{time interval of data from startdate (hours)}

\item{my.startdate}{start date of the requested data must be given in ISO 8601 Format (https://de.wikipedia.org/wiki/ISO_8601) e.g. "2002-12-12T12:00Z"}

\item{aggregation}{how data shall be agregated ("MEAN","SUM")}
}
\value{
a json file with all available sensors
}
\description{
Downloads hourly data for a given sensor provided by OpenSensorWeb REST API (https://www.opensensorweb.de/).

OpenSensorWeb provides data in UTC-Timeformat. If TimeExtent is providet in the Request in UTC format (2002-12-12T12:00Z) data is returned in UTC format. Requests must use ISO 8601 Format: https://de.wikipedia.org/wiki/ISO_8601 while the "Z" in 2002-12-12T12:00Z notates for UTC and is equivalent to 2002-12-12T12:00+00:00
requests might also be defined using intervals (e.g. 2002-09-12T12:00Z/2002-12-12T12:00Z/PT1H (time-based multi-interval, UTC)) but this is not implemented in this package.
}
\examples{
hourly(url = "https://api.sensoto.io/v1/organizations/open/networks/AMMS_WETTERDATEN", my.device = "S021", my.sensor = "Niederschlag", aggregation = "MEAN", my.interval = 1000, my.startdate = "2015-09-18T00:00:00Z")
}
